Separated out SocketIO app and moved zipping to webdata_aggregator.
